Isolate move FUZZ_TEST and MemoryManager() out of test_utils (#3617)
(cherry picked from commit c8e435cbe70fac9679d05367e8297c0a27aacd13)
