+++ lib/extras/test_memory_manager.h
@@ -1,0 +1,19 @@
+// Copyright (c) the JPEG XL Project Authors. All rights reserved.
+//
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+#ifndef LIB_EXTRAS_TEST_MEMORY_MANAGER_H_
+#define LIB_EXTRAS_TEST_MEMORY_MANAGER_H_
+
+#include "lib/base/memory_manager.h"
+
+namespace jxl {
+namespace test {
+
+JxlMemoryManager* MemoryManager();
+
+}  // namespace test
+}  // namespace jxl
+
+#endif  // LIB_EXTRAS_TEST_MEMORY_MANAGER_H_
