+++ lib/lib.gni
@@ -522,6 +522,7 @@ libjxl_jpegli_sources = [
 ]
 
 libjxl_jpegli_testlib_files = [
+    "jpegli/fuzztest.h",
     "jpegli/test_params.h",
     "jpegli/test_utils-inl.h",
     "jpegli/test_utils.cc",
@@ -573,6 +574,8 @@ libjxl_testlib_files = [
     "jxl/render_pipeline/test_render_pipeline_stages.h",
     "jxl/test_image.cc",
     "jxl/test_image.h",
+    "jxl/test_memory_manager.cc",
+    "jxl/test_memory_manager.h",
     "jxl/test_utils.cc",
     "jxl/test_utils.h",
 ]
